array와 list
