+++ app/src/main/java/com/example/test/main.kt
@@ -1,0 +1,16 @@
+package com.example.test
+
+fun main() {
+    helloWorld()
+
+    add(2, 4)
+
+    stringTemplate()
+
+    checkNum(2)
+}
+//Expression vs Statement
+// Expression : return 값이 있는 것
+// Statement : return 값이 없는 것
+// 코틀린에서 모든 함수는 Expression이다. -> return값이 없는 것도 사실은 Unit을 반환함
+
